Worked on Teleporter Script
Worked on a handful of things. Created scripts for future/current writing. Updated test_scene_01 with the Teleportation Script and a quickly thrown together asset. Worked on a good bit of the Teleporter Script.

Currently the Teleporter Script does not work (attempting to "teleport" (swap scenes) results in an error that doesn't crash the game, but does nothing).