Chain UI Scrolling
Implemented the functionality to scroll up and down through the chain, as well as have it reset to the most recently played card from any current position in the chain.

Currently, it is lacking a UI functionality as it relies solely on the L and . keys (Scroll upwards, Scroll downwards). There is also the issue of the cards instantly disappearing and snapping when it performs a Reset Scroll to most Recently Played Card.