Changed fSwitchCardPosX Values back to 4.9
Values were changed back to their previous amount of 4.9

Although 4.1 might work, it would require changing the battleground prefab to have it work along with a few new manipulations of the variable (it wouldn't be too hard though, just a quick find and replace).

If this is not good or against the desired, let me know and I can do a quick revert and have the new fixes line up with a 4.1f value.