Updates to TextUpdates
In textupdates, I changed some names for better readability. I changed update text to be called when it is needed, so it no longer runs 60 times a frame. I removed some code as a result of that and found a new error involving the UI not hiding after the card is played for the first time. I fixed that by making a Coroutine that waits a second after the card is played, so it overrides whatever is going on.