Implemented changes/fixes for the Sprint Build
Updated prefabs, fixed a few coding issues, and attempted to hotfix the Health and Defense Bars. Currently there are still some documented issues present and features not properly showcased in this build, but this does not limit the possibility of there being another update today or this weekend that addresses some of these issues.