Attempted to Fix Health and Defense Bars
Implemented what seems like a fix to Health and Defense, but after hours of struggling, I decided to leave it here and return later. I have other tasks that need to be worked on in the meantime, so if there are issues, please mark them down so that I may review them later and attempt to patch in those fixes!