Enemy not playing cards when they have them.
Fixed the issue of enemies not playing cards sometimes when they're able to. Couldn't narrow it to one specific issue after a day of trying, so I removed a small line of code (|| failProb > maxFailProb) for now. The enemy now plays cards when they have them no matter what, and it keeps the game moving at a steady pace. I also fixed an issue with the enemy hand not replenishing when they end turn entirely.