Temporarily Fixed Tutorial Issue
Had to comment out three lines of code in order to have the Tutorial Work properly. Section has been marked accordingly to help find what needs to be addressed. Was going to fix it myself, but I did not want to spend a large amount of time in case there are more game breaking bugs.