Solved Tutorial Issue in Lose Scene Issue
Found the source of the Tutorial Scene bug. It was related to the Card Count in the deck, so I changed it from 6 to 30 (the default).

Fixed Lose Scene not activating Main Menu's music track properly because the Audio Manager prefab for that scene in particular did not update when updating the source prefab directly, which is why the bug slipped in without notice.